Работа с кнопкой Add review в MoviePageScreen - обзорную секцию фильма вынесет в отдельный компонент - выполнит проверку для кнопки Add review, будет показана для авториз.
