+++ src/components/movie-page-screen/movie-page-screen.jsx
@@ -1,27 +1,24 @@
 import React, {useEffect} from 'react';
-import {useHistory} from 'react-router-dom';
 import PropTypes from "prop-types";
-import MoviesList from '../movies-list/movies-list';
-import {propFilm} from '../../props-validation';
-import Tabs from './tabs/tabs';
-import {getSimilarFilms} from '../../utils';
-import {Lists} from '../../const';
-import {fetchFilm} from '../../store/api-actions';
 import {connect} from 'react-redux';
+import {fetchFilm} from '../../store/api-actions';
+import {Lists} from '../../const';
+import {getSimilarFilms} from '../../utils';
+import {propFilm} from '../../props-validation';
+
+import MovieCardFull from './movie-card-full/movie-card-full';
+import Tabs from './tabs/tabs';
+import MoviesList from '../movies-list/movies-list';
 import LoadingScreen from '../loading-screen/loading-screen';
-import Header from '../header/header';
-import Footer from '../footer/footer';
 import NotFoundScreen from '../not-found-screen/not-found-screen';
+import Footer from '../footer/footer';
 
-const MoviePageScreen = ({films, film, id, isLoadFilm, onLoadFilm, isLoadFilmFailed}) => {
-  const history = useHistory();
-
+const MoviePageScreen = ({films, film, id, isLoadFilm, onLoadFilm, isLoadFilmFailed, authorizationStatus}) => {
   useEffect(() => {
     if (!isLoadFilm) {
       onLoadFilm(id);
     }
   }, [isLoadFilm]);
-
 
   if (!isLoadFilm && !isLoadFilmFailed) {
     return <LoadingScreen />;
@@ -29,58 +26,22 @@ const MoviePageScreen = ({films, film, id, isLoadFilm, onLoadFilm, isLoadFilmFai
     return <NotFoundScreen />;
   }
 
-  const {backgroundImage, backgroundColor, name, genre, released, posterImage} = film;
   const similarFilms = getSimilarFilms(films, film);
 
   return (
     <>
-      <section className="movie-card movie-card--full" style={{backgroundColor}}>
-        <div className="movie-card__hero">
-          <div className="movie-card__bg">
-            <img src={backgroundImage} alt={name} />
-          </div>
-          <h1 className="visually-hidden">WTW</h1>
-          <Header />
-          <div className="movie-card__wrap">
-            <div className="movie-card__desc">
-              <h2 className="movie-card__title">{name}</h2>
-              <p className="movie-card__meta">
-                <span className="movie-card__genre">{genre}</span>
-                <span className="movie-card__year">{released}</span>
-              </p>
-              <div className="movie-card__buttons">
-                <button className="btn btn--play movie-card__button" type="button" onClick={() => history.push(`/player/${id}`)}>
-                  <svg viewBox="0 0 19 19" width={19} height={19}>
-                    <use xlinkHref="#play-s" />
-                  </svg>
-                  <span>Play</span>
-                </button>
-                <button className="btn btn--list movie-card__button" type="button">
-                  <svg viewBox="0 0 19 20" width={19} height={20}>
-                    <use xlinkHref="#add" />
-                  </svg>
-                  <span>My list</span>
-                </button>
-                <a href="add-review.html" className="btn movie-card__button">Add review</a>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div className="movie-card__wrap movie-card__translate-top">
-          <div className="movie-card__info">
-            <div className="movie-card__poster movie-card__poster--big">
-              <img src={posterImage} alt={name} width={218} height={327} />
-            </div>
-            <Tabs film={film}/>
-          </div>
-        </div>
-      </section>
+      <MovieCardFull film={film} id={id} authorizationStatus={authorizationStatus}>
+        <Tabs film={film}/>
+      </MovieCardFull>
+
       <div className="page-content">
         <section className="catalog catalog--like-this">
           <h2 className="catalog__title">More like this</h2>
           {similarFilms && <MoviesList films={similarFilms.splice(0, Lists.MAX_SIMILAR)} />}
         </section>
+
         <Footer />
+
       </div>
     </>
   );
@@ -95,6 +56,7 @@ MoviePageScreen.propTypes = {
   isLoadFilm: PropTypes.bool.isRequired,
   onLoadFilm: PropTypes.func.isRequired,
   isLoadFilmFailed: PropTypes.bool.isRequired,
+  authorizationStatus: PropTypes.string.isRequired,
 };
 
 const mapStateToProps = (state) => {
@@ -102,6 +64,7 @@ const mapStateToProps = (state) => {
     film: state.film,
     isLoadFilm: state.isLoadFilm,
     isLoadFilmFailed: state.isLoadFilmFailed,
+    authorizationStatus: state.authorizationStatus,
   };
 };
